User commit

+++ src/main/java/com/example/demo/service/UserService.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
-import com.example.demo.entity.User;
+import com.example.demo.entity.UserEntity;
 import com.example.demo.repository.UserRepo;
 
 @Service
@@ -17,23 +17,39 @@ public class UserService {
         this.userRepo = userRepo;
     }
 
-    // CREATE / UPDATE
-    public User insertUser(User user) {
+    // CREATE
+    public UserEntity insertUser(UserEntity user) {
         return userRepo.save(user);
     }
 
     // READ ALL
-    public List<User> getAllUsers() {
+    public List<UserEntity> getAllUsers() {
         return userRepo.findAll();
     }
 
     // READ ONE
-    public Optional<User> getOneUser(Long id) {
+    public Optional<UserEntity> getOneUser(Long id) {
         return userRepo.findById(id);
     }
 
+    // UPDATE
+    public UserEntity updateUser(Long id, UserEntity newUser) {
+        return userRepo.findById(id)
+            .map(user -> {
+                user.setName(newUser.getName());
+                user.setEmail(newUser.getEmail());
+                user.setCgpa(newUser.getCgpa());
+                user.setDob(newUser.getDob());
+                return userRepo.save(user);
+            }).orElse(null);
+    }
+
     // DELETE
-    public void deleteUser(Long id) {
-        userRepo.deleteById(id);
+    public boolean deleteUser(Long id) {
+        if (userRepo.existsById(id)) {
+            userRepo.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
